Generate IRs of Cases entities in Psc::ScheduledActivityReport.  #2255.
