+++ spec/models/psc/scheduled_activity_report/entity_resolution_spec.rb
@@ -1,87 +1,19 @@
 require 'spec_helper'
-
 require 'set'
 
-class Psc::ScheduledActivityReport
-  describe EntityMapping do
-    EM = EntityMapping
+require File.expand_path('../../example_data', __FILE__)
 
-    let(:data) { JSON.parse(File.read(File.expand_path('../../ex1.json', __FILE__))) }
+class Psc::ScheduledActivityReport
+  describe EntityResolution do
+    include_context 'example data'
+
     let(:report) { ::Psc::ScheduledActivityReport.from_json(data) }
 
-    # These values are derived from ex1.json.
-    let(:activity_name) { 'Low-Intensity Interview' }
-    let(:event_data_collection) { 'event:low_intensity_data_collection' }
-    let(:event_informed_consent) { 'event:informed_consent' }
-    let(:ideal_date) { '2012-07-06' }
-    let(:instrument_pregnotpreg) { 'instrument:ins_que_lipregnotpreg_int_li_p2_v2.0' }
-    let(:person_id) { '2f85c94e-edbb-4cbe-b9ab-5f12c033323f' }
-    let(:scheduled_date) { '2012-07-10' }
-
     before do
-      report.extend(EntityMapping)
+      report.extend(EntityResolution)
     end
 
     describe '#process' do
-      before do
-        report.process
-      end
-
-      it 'finds people' do
-        report.people.should == Set.new([EM::Person.new(person_id)])
-      end
-
-      it 'finds contacts' do
-        person = EM::Person.new(person_id)
-
-        report.contacts.should == Set.new([EM::Contact.new(scheduled_date, person)])
-      end
-
-      it 'finds surveys' do
-        report.surveys.should == Set.new([
-          EM::Survey.new(instrument_pregnotpreg)
-        ])
-      end
-
-      it 'finds events' do
-        person = EM::Person.new(person_id)
-        contact = EM::Contact.new(scheduled_date, person)
-
-        report.events.should == Set.new([
-          EM::Event.new(event_informed_consent, ideal_date, contact, person),
-          EM::Event.new(event_data_collection, ideal_date, contact, person)
-        ])
-      end
-
-      it 'finds instruments' do
-        person = EM::Person.new(person_id)
-        contact = EM::Contact.new(scheduled_date, person)
-        event = EM::Event.new(event_data_collection, ideal_date, contact, person)
-        survey = EM::Survey.new(instrument_pregnotpreg)
-
-        report.instruments.should == Set.new([
-          EM::Instrument.new(survey, activity_name, event, person)
-        ])
-      end
-
-      # NOTE: This isn't actually testing for ContactLinks.  It, like the rest
-      # of the examples, is testing for ContactLink precursors.
-      it 'links a person, contact, event, and instrument' do
-        person = EM::Person.new(person_id)
-        contact = EM::Contact.new(scheduled_date, person)
-        event1 = EM::Event.new(event_informed_consent, ideal_date, contact, person)
-        event2 = EM::Event.new(event_data_collection, ideal_date, contact, person)
-        survey = EM::Survey.new(instrument_pregnotpreg)
-        instrument = EM::Instrument.new(survey, activity_name, event2, person)
-
-        report.contact_links.should == Set.new([
-          EM::ContactLink.new(person, contact, event1, nil),
-          EM::ContactLink.new(person, contact, event2, instrument)
-        ])
-      end
-    end
-
-    describe '#resolve_models' do
       let(:sio) { StringIO.new }
       let(:log) { sio.string }
       let(:staff_id) { 'fa542082-c96f-4886-a6bc-cc9a546d787a' }
@@ -89,8 +21,6 @@ class Psc::ScheduledActivityReport
       before do
         report.logger = ::Logger.new(sio)
         report.staff_id = staff_id
-
-        report.process
       end
 
       describe 'with #staff_id blank' do
@@ -99,7 +29,7 @@ class Psc::ScheduledActivityReport
         end
 
         it 'raises an error' do
-          lambda { report.resolve_models }.should raise_error
+          lambda { report.process }.should raise_error
         end
       end
 
@@ -107,13 +37,13 @@ class Psc::ScheduledActivityReport
         it 'finds people in Cases' do
           p = Factory(:person, :person_id => person_id)
 
-          report.resolve_models
+          report.process
 
           report.people.models.should == Set.new([p])
         end
 
         it 'logs an error if a person cannot be found' do
-          report.resolve_models
+          report.process
 
           log.should =~ /cannot map {person ID = #{person_id}} to a person/i
         end
@@ -123,13 +53,13 @@ class Psc::ScheduledActivityReport
         it 'finds surveys in Cases' do
           s = Factory(:survey, :access_code => 'ins_que_lipregnotpreg_int_li_p2')
 
-          report.resolve_models
+          report.process
 
           report.surveys.models.should == Set.new([s])
         end
 
         it 'logs an error if a survey cannot be found' do
-          report.resolve_models
+          report.process
 
           log.should =~ /cannot map {access code = ins_que_lipregnotpreg_int_li_p2} to a survey/i
         end
@@ -152,13 +82,13 @@ class Psc::ScheduledActivityReport
           et2 = NcsCode.for_list_name_and_local_code('EVENT_TYPE_CL1', 10)
           e2 = Factory(:event, :participant => pa, :event_start_date => ideal_date, :event_type => et2)
 
-          report.resolve_models
+          report.process
 
           report.events.models.should == Set.new([e1, e2])
         end
 
         it 'logs an error if an event cannot be found' do
-          report.resolve_models
+          report.process
 
           log.should =~ /cannot map {label = event:low_intensity_data_collection, ideal date = #{ideal_date}, participant = #{pa.p_id}} to an event/i
           log.should =~ /cannot map {label = event:informed_consent, ideal date = #{ideal_date}, participant = #{pa.p_id}} to an event/i
@@ -187,7 +117,7 @@ class Psc::ScheduledActivityReport
           let!(:s) { Factory(:survey, :access_code => 'ins_que_lipregnotpreg_int_li_p2', :title => instrument_pregnotpreg) }
 
           it 'starts an instrument' do
-            report.resolve_models
+            report.process
 
             report.instruments.models.first.should_not be_nil
           end
@@ -196,7 +126,7 @@ class Psc::ScheduledActivityReport
             let(:instrument) { report.instruments.models.first }
 
             before do
-              report.resolve_models
+              report.process
             end
 
             it 'is a new record' do
@@ -229,7 +159,7 @@ class Psc::ScheduledActivityReport
           end
 
           it 'reuses that contact' do
-            report.resolve_models
+            report.process
 
             report.contacts.models.should == Set.new([c])
           end
@@ -237,7 +167,7 @@ class Psc::ScheduledActivityReport
 
         describe 'if there does not exist a contact for the scheduled date, person, and staff ID' do
           it 'starts a new contact' do
-            report.resolve_models
+            report.process
 
             report.contacts.models.first.should_not be_nil
           end
@@ -246,7 +176,7 @@ class Psc::ScheduledActivityReport
             let(:contact) { report.contacts.models.first }
 
             before do
-              report.resolve_models
+              report.process
             end
 
             it 'is a new record' do
@@ -262,7 +192,7 @@ class Psc::ScheduledActivityReport
 
           let!(:s) { Factory(:survey, :access_code => 'ins_que_lipregnotpreg_int_li_p2', :title => instrument_pregnotpreg) }
           let!(:c) { Factory(:contact, :contact_date => scheduled_date) }
-          let!(:i) { Instrument.start(p, s, e).tap(&:save!) }
+          let!(:i) { ::Instrument.start(p, s, e).tap(&:save!) }
 
           describe 'if a link already exists' do
             let!(:cl) do
@@ -271,7 +201,7 @@ class Psc::ScheduledActivityReport
             end
 
             it 'reuses that link' do
-              report.resolve_models
+              report.process
 
               report.contact_links.models.should include(cl)
             end
@@ -279,7 +209,7 @@ class Psc::ScheduledActivityReport
 
           describe 'if a link does not exist' do
             it 'builds links' do
-              report.resolve_models
+              report.process
 
               report.contact_links.models.none?(&:nil?).should be_true
             end
@@ -288,7 +218,7 @@ class Psc::ScheduledActivityReport
               let(:links) { report.contact_links.models }
 
               before do
-                report.resolve_models
+                report.process
               end
 
               it 'contains staff ID' do
@@ -333,7 +263,6 @@ class Psc::ScheduledActivityReport
         let!(:e2) { Factory(:event, :participant => pa, :event_start_date => ideal_date, :event_type => et2) }
         let!(:s) { Factory(:survey, :access_code => 'ins_que_lipregnotpreg_int_li_p2', :title => instrument_pregnotpreg) }
 
-
         before do
           # Link up.
           p.participant = pa
@@ -342,7 +271,6 @@ class Psc::ScheduledActivityReport
           report.logger = ::Logger.new(sio)
           report.staff_id = staff_id
           report.process
-          report.resolve_models
 
           log.should be_empty
